be_rtos/requirements.txt
django==5.0.6
mysqlclient==2.2.4
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
django-cors-headers==4.3.1
django-guardian==2.4.0
drf-spectacular-sidecar==2024.6.1
drf-standardized-errors==0.13.0
pika==1.3.2
firebase-admin==6.5.0
django-extensions==3.2.3
mqttasgi==1.2.4
paho-mqtt==2.1.0
django-shortuuidfield==0.1.3